Codechange: Move dropdown and slider out of widgets directory.
Also shuffle headers to place widget includes near end.

This leaves the widgets directory solely for defining Widget IDs.
