+++ src/CMakeLists.txt
@@ -138,6 +138,9 @@ add_files(
     dock_gui.cpp
     driver.cpp
     driver.h
+    dropdown.cpp
+    dropdown_func.h
+    dropdown_type.h
     economy.cpp
     economy_base.h
     economy_cmd.h
@@ -393,6 +396,8 @@ add_files(
     signs_func.h
     signs_gui.cpp
     signs_type.h
+    slider.cpp
+    slider_func.h
     slope_func.h
     slope_type.h
     smallmap_gui.cpp
