+++ src/build_vehicle_gui.cpp
@@ -25,7 +25,7 @@
 #include "window_func.h"
 #include "timer/timer_game_calendar.h"
 #include "vehicle_func.h"
-#include "widgets/dropdown_func.h"
+#include "dropdown_func.h"
 #include "engine_gui.h"
 #include "cargotype.h"
 #include "core/geometry_func.hpp"
