+++ src/game/game_gui.cpp
@@ -14,7 +14,7 @@
 #include "../window_func.h"
 #include "../network/network.h"
 #include "../network/network_content.h"
-#include "../widgets/dropdown_func.h"
+#include "../dropdown_func.h"
 #include "../timer/timer.h"
 #include "../timer/timer_window.h"
