+++ src/genworld_gui.cpp
@@ -18,8 +18,8 @@
 #include "sound_func.h"
 #include "fios.h"
 #include "string_func.h"
-#include "widgets/dropdown_type.h"
-#include "widgets/dropdown_func.h"
+#include "dropdown_type.h"
+#include "dropdown_func.h"
 #include "querystring_gui.h"
 #include "town.h"
 #include "core/geometry_func.hpp"
