+++ src/newgrf_gui.cpp
@@ -16,8 +16,8 @@
 #include "gamelog.h"
 #include "settings_type.h"
 #include "settings_func.h"
-#include "widgets/dropdown_type.h"
-#include "widgets/dropdown_func.h"
+#include "dropdown_type.h"
+#include "dropdown_func.h"
 #include "network/network.h"
 #include "network/network_content.h"
 #include "sortlist_type.h"
