+++ src/rail_gui.h
@@ -11,7 +11,7 @@
 #define RAIL_GUI_H
 
 #include "rail_type.h"
-#include "widgets/dropdown_type.h"
+#include "dropdown_type.h"
 
 struct Window *ShowBuildRailToolbar(RailType railtype);
 void ReinitGuiAfterToggleElrail(bool disable);
