+++ src/screenshot_gui.cpp
@@ -11,9 +11,13 @@
 #include "window_func.h"
 #include "window_gui.h"
 #include "screenshot.h"
-#include "widgets/screenshot_widget.h"
-#include "table/strings.h"
 #include "gfx_func.h"
+
+#include "widgets/screenshot_widget.h"
+
+#include "table/strings.h"
+
+#include "safeguards.h"
 
 struct ScreenshotWindow : Window {
 	ScreenshotWindow(WindowDesc *desc) : Window(desc)
