+++ src/settings_gui.cpp
@@ -19,9 +19,9 @@
 #include "strings_func.h"
 #include "window_func.h"
 #include "string_func.h"
-#include "widgets/dropdown_type.h"
-#include "widgets/dropdown_func.h"
-#include "widgets/slider_func.h"
+#include "dropdown_type.h"
+#include "dropdown_func.h"
+#include "slider_func.h"
 #include "highscore.h"
 #include "base_media_base.h"
 #include "company_base.h"
