+++ src/smallmap_gui.cpp
@@ -26,10 +26,11 @@
 #include "strings_func.h"
 #include "blitter/factory.hpp"
 #include "linkgraph/linkgraph_gui.h"
-#include "widgets/smallmap_widget.h"
 #include "timer/timer.h"
 #include "timer/timer_window.h"
 #include "smallmap_gui.h"
+
+#include "widgets/smallmap_widget.h"
 
 #include "table/strings.h"
