+++ src/station_cmd.cpp
@@ -55,7 +55,6 @@
 #include "company_gui.h"
 #include "linkgraph/linkgraph_base.h"
 #include "linkgraph/refresh.h"
-#include "widgets/station_widget.h"
 #include "tunnelbridge_map.h"
 #include "station_cmd.h"
 #include "waypoint_cmd.h"
@@ -67,6 +66,8 @@
 #include "timer/timer_game_economy.h"
 #include "timer/timer_game_tick.h"
 #include "cheat_type.h"
+
+#include "widgets/station_widget.h"
 
 #include "table/strings.h"
