+++ src/vehicle_gui_base.h
@@ -17,7 +17,7 @@
 #include "vehicle_base.h"
 #include "vehiclelist.h"
 #include "window_gui.h"
-#include "widgets/dropdown_type.h"
+#include "dropdown_type.h"
 
 typedef GUIList<const Vehicle*, std::nullptr_t, CargoID> GUIVehicleList;
