+++ src/widgets/CMakeLists.txt
@@ -11,9 +11,6 @@ add_files(
     date_widget.h
     depot_widget.h
     dock_widget.h
-    dropdown.cpp
-    dropdown_func.h
-    dropdown_type.h
     dropdown_widget.h
     engine_widget.h
     error_widget.h
@@ -48,8 +45,6 @@ add_files(
     script_widget.h
     settings_widget.h
     sign_widget.h
-    slider.cpp
-    slider_func.h
     smallmap_widget.h
     station_widget.h
     statusbar_widget.h
